fix: fix comparison of integer expressions of different signedness 'int' and 'uint64_t'

+++ src/cpu/cpu-exec.c
@@ -83,7 +83,7 @@ static void update_instr_cnt() {
   IFNDEF(CONFIG_DEBUG, g_nr_guest_instr += n_executed);
 
   n_remain =
-      n_batch > n_remain_total ? n_remain_total : n_batch; // clean n_remain
+      n_batch > (int)n_remain_total ? n_remain_total : n_batch; // clean n_remain
   // Loge("n_remain = %i, n_remain_total = %lu\n", n_remain, n_remain_total);
 #endif
 }
